Refactor to not include crates
