--- crates/karva_core/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "karva_core"
-version = "0.0.0"
-edition = "2024"
-
-[lib]
-crate-type = ["rlib", "cdylib"]
-
-[dependencies]
-camino = { workspace = true }
-colored = { workspace = true }
-pyo3 = { version = "0.25.0", features = ["extension-module"] }
-rustpython-parser = { version = "0.4.0" }
-serde = { workspace = true }
-tracing = { workspace = true }
-tracing-subscriber = { workspace = true, optional = true }
-tracing-tree = { workspace = true, optional = true }
-ignore = "0.4.23"
-
-[dev-dependencies]
-tempfile = "3.8"
-regex = "1"
-
-[features]
-# must be enabled when building with `cargo build`, maturin enables this automatically
-extension-module = ["pyo3/extension-module"]
