+++ src/cli/logging.rs
