+++ src/cli/mod.rs
@@ -1,0 +1,4 @@
+mod args;
+mod logging;
+pub mod main;
+mod version;
