+++ src/core/project.rs
@@ -1,4 +1,4 @@
-use crate::path::{PythonTestPath, SystemPathBuf};
+use crate::core::path::{PythonTestPath, SystemPathBuf};
 
 pub struct Project {
     cwd: SystemPathBuf,
