+++ src/main.rs
@@ -1,0 +1,5 @@
+use karva::cli::main::main as cli_main;
+
+fn main() {
+    cli_main();
+}
